pyproject.toml
[project]
name = "detent"
version = "1.0.0"
description = "A verification runtime that intercepts AI coding agent file writes, runs them through a configurable verification pipeline, and rolls back atomically on failure."
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    { name = "Ofir Cohen", email = "ofircohen205@gmail.com" },
]
keywords = ["ai", "coding-agent", "verification", "rollback", "lint", "typecheck"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "aiohttp>=3.9,<4",
    "certifi>=2024.0",
    "pyyaml>=6.0,<7",
    "pydantic>=2.0,<3",
    "tree-sitter>=0.23,<1",
    "tree-sitter-python>=0.23,<1",
    "tree-sitter-javascript>=0.23,<1",
    "tree-sitter-typescript>=0.23,<1",
    "tree-sitter-go>=0.23,<1",
    "tree-sitter-rust>=0.23,<1",
    "click>=8.1,<9",
    "rich>=13.7,<14",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=6.0,<7",
    "pip-audit>=2.7,<3",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
    "mypy-extensions>=1.0",
    "types-PyYAML>=6.0",
    "twine>=6.1.0",
]
security = [
    "bandit>=1.9.4",
    "pip-audit>=2.7,<3",
    "semgrep>=1.60",
]
telemetry = [
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]

[project.scripts]
detent = "detent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ─── Tool Configuration ──────────────────────────────────────────────────────

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "opentelemetry.exporter.otlp.proto.grpc"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: fast unit tests (no external deps)",
    "integration: full pipeline tests (requires tools installed)",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "types-tree-sitter>=0.20.1.20240311",
    "pip-audit>=2.7,<3",
]
